MODIFIED: move all examples to src/examples/ folder - Fixed #62
