--- examples/java/semWeb/run.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-. ../env.sh
-
-run_preloaded SemWeb
-
