--- examples/java/test3/run.bat
@@ -1,11 +1,0 @@
-@echo off
-call ..\env.bat
-
-if not exist Test.class (
-  echo  Compiling Test.java
-  javac Test.java
-)
-
-java Test
-
-pause
