--- examples/java/test3/run.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-. ../env.sh
-
-run Test
-
