--- examples/java/testGC/run.bat
@@ -1,11 +1,0 @@
-@echo off
-call ..\env.bat
-
-if not exist TestGC.class (
-  echo  Compiling TestGC.java
-  javac TestGC.java
-)
-
-java TestGC 100000
-
-pause
