--- examples/java/testGC/run.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-. ../env.sh
-
-run TestGC $*
-
