+++ src/examples/java/semWeb/SemWeb.java
