+++ src/examples/java/semWeb/test.rdf
