+++ src/examples/java/sources.txt
@@ -1,8 +1,12 @@
-./org/jpl7/examples/family/FamilyMT.java
-./org/jpl7/examples/family/Family.java
-./org/jpl7/examples/zahed/Zahed.java
-./org/jpl7/examples/factorial/Factorial.java
-./org/jpl7/examples/Exceptions.java
-./org/jpl7/examples/time/Time.java
-./org/jpl7/examples/test/Test.java
-./org/jpl7/examples/Versions.java
+./family/FamilyMT.java
+./family/Family.java
+./zahed/Zahed.java
+./testGC/TestGC.java
+./factorial/Factorial.java
+./semWeb/SemWeb.java
+./time/Time.java
+./test/Test.java
+./db/Db.java
+./system/Exceptions.java
+./system/Init.java
+./system/Versions.java
