+++ src/examples/java/system/Versions.java
@@ -1,4 +1,4 @@
-package org.jpl7.examples;
+package system;
 
 import org.jpl7.JPL;
 import org.jpl7.Query;
