+++ src/examples/java/testGC/TestGC.java
