+++ src/examples/java/zahed/run.bat
