+++ src/examples/java/zahed/run.sh
