+++ src/examples/java/zahed/zahed.pl
