+++ src/examples/prolog/jpl_colour_choose_demo.pl
