+++ src/examples/prolog/jpl_jlist_demo.pl
