+++ src/examples/prolog/jpl_table_demo.pl
