+++ src/examples/prolog/jpl_text_entry_demo.pl
